# Improve `ClassDB::_is_parent_class` performance