Change `ClassDB::_is_parent_class` to use `ClassInfo::inherits_ptr`, instead of looking up each inherited class name.

This speeds up `Array` type checking - adding to a typed `Array` can be around 2x faster, depending on how many inherited classes are between the object being added and the `Array`'s type.  Also speeds up casting/type checking in gdscript with `as`/`is` for native types.

Comparison gdscript below:

```gdscript
func _ready() -> void:
	test_untyped_array()
	test_node_button_array()
	test_node_node3d_array()
	test_cast()

func test_untyped_array():
	var a : Array = []
	var start_time = Time.get_ticks_msec()
	for i in 1000000:
		a.push_back(i)
	var end_time = Time.get_ticks_msec()
	print("test_untyped_array: %dms" % (end_time - start_time))

func test_node_button_array():
	var a : Array[Node] = []
	var b := Button.new()
	var start_time = Time.get_ticks_msec()
	for i in 1000000:
		a.push_back(b)
	var end_time = Time.get_ticks_msec()
	print("test_node_button_array: %dms" % (end_time - start_time))

func test_node_node3d_array():
	var a : Array[Node] = []
	var n := Node3D.new()
	var start_time = Time.get_ticks_msec()
	for i in 1000000:
		a.push_back(n)
	var end_time = Time.get_ticks_msec()
	print("test_node_node3d_array: %dms" % (end_time - start_time))
	
func test_cast():
	var event : InputEvent = InputEventMouseMotion.new()
	var start_time = Time.get_ticks_msec()
	for i in 1000000:
		event as InputEventKey
	var end_time = Time.get_ticks_msec()
	print("test_cast: %dms" % (end_time - start_time))
```

Old
```
test_untyped_array: 89ms
test_node_button_array: 402ms
test_node_node3d_array: 242ms
test_cast: 508ms
```

New
```
test_untyped_array: 88ms
test_node_button_array: 155ms
test_node_node3d_array: 155ms
test_cast: 88ms
```


<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->